requirements.txt
beautifulsoup4==4.10.0
pandas==1.3.4
python-dotenv==0.19.2
selenium==4.1.3
webdriver-manager==3.5.4